Flytt funksjoner til utils-fil
