+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/beregning/SmåbarnstilleggBarnetrygdGenerator.kt
@@ -4,7 +4,6 @@ import no.nav.familie.ba.sak.common.DatoIntervallEntitet
 import no.nav.familie.ba.sak.common.Feil
 import no.nav.familie.ba.sak.common.MånedPeriode
 import no.nav.familie.ba.sak.common.Utils.avrundetHeltallAvProsent
-import no.nav.familie.ba.sak.common.erUnder3ÅrTidslinje
 import no.nav.familie.ba.sak.common.førsteDagIInneværendeMåned
 import no.nav.familie.ba.sak.common.sisteDagIInneværendeMåned
 import no.nav.familie.ba.sak.common.sisteDagIMåned
@@ -19,10 +18,7 @@ import no.nav.familie.ba.sak.kjerne.beregning.domene.TilkjentYtelse
 import no.nav.familie.ba.sak.kjerne.beregning.domene.YtelseType
 import no.nav.familie.ba.sak.kjerne.personident.Aktør
 import no.nav.familie.ba.sak.kjerne.tidslinje.Tidslinje
-import no.nav.familie.ba.sak.kjerne.tidslinje.komposisjon.kombinerMed
-import no.nav.familie.ba.sak.kjerne.tidslinje.komposisjon.kombinerUtenNull
 import no.nav.familie.ba.sak.kjerne.tidslinje.tid.Måned
-import no.nav.familie.ba.sak.kjerne.tidslinje.transformasjon.beskjærEtter
 import no.nav.familie.ba.sak.kjerne.tidslinje.transformasjon.tilMåned
 import no.nav.fpsak.tidsserie.LocalDateSegment
 import no.nav.fpsak.tidsserie.LocalDateTimeline
@@ -62,25 +58,6 @@ data class SmåbarnstilleggBarnetrygdGenerator(
             søkerAktør = utvidetAndeler.first { it.erUtvidet() }.aktør
         )
     }
-
-    private fun kombinerAlleTidslinjerTilProsentTidslinje(
-        perioderMedFullOvergangsstønadTidslinje: Tidslinje<Boolean, Måned>,
-        utvidetBarnetrygdTidslinje: AndelTilkjentYtelseTidslinje,
-        barnSomGirRettTilSmåbarnstilleggTidslinje: Tidslinje<BarnSinRettTilSmåbarnstillegg, Måned>
-    ) =
-        perioderMedFullOvergangsstønadTidslinje
-            .kombinerMed(utvidetBarnetrygdTidslinje) { overgangsstønadTidslinje, utvidetTidslinje ->
-                if (overgangsstønadTidslinje == null || utvidetTidslinje == null) null
-                else if (utvidetTidslinje.prosent > BigDecimal.ZERO) UtvidetAndelStatus.UTBETALING
-                else UtvidetAndelStatus.NULLUTBETALING
-            }
-            .kombinerMed(barnSomGirRettTilSmåbarnstilleggTidslinje) { overgangsstønadOgUtvidetTidslinje, under3ÅrTidslinje ->
-                if (overgangsstønadOgUtvidetTidslinje == null || under3ÅrTidslinje == null) null
-                else if (under3ÅrTidslinje == BarnSinRettTilSmåbarnstillegg.UNDER_3_ÅR_UTBETALING && overgangsstønadOgUtvidetTidslinje == UtvidetAndelStatus.UTBETALING) BigDecimal(
-                    100
-                )
-                else BigDecimal.ZERO
-            }
 
     fun lagSmåbarnstilleggAndelerGammel(
         perioderMedFullOvergangsstønad: List<InternPeriodeOvergangsstønad>,
@@ -196,35 +173,6 @@ data class SmåbarnstilleggBarnetrygdGenerator(
                 prosent = prosentIPeriode
             )
         }
-    }
-
-    private fun kombinerBarnasTidslinjerTilUnder3ÅrResultat(
-        alleAndelerForBarnUnder3År: Iterable<AndelTilkjentYtelse>
-    ): BarnSinRettTilSmåbarnstillegg? {
-        val høyesteProsentIPeriode = alleAndelerForBarnUnder3År.maxOfOrNull { it.prosent }
-
-        return when (høyesteProsentIPeriode) {
-            null -> null
-            BigDecimal.ZERO -> BarnSinRettTilSmåbarnstillegg.UNDER_3_ÅR_NULLUTBETALING
-            else -> BarnSinRettTilSmåbarnstillegg.UNDER_3_ÅR_UTBETALING
-        }
-    }
-
-    fun lagTidslinjeForPerioderMedBarnSomGirRettTilSmåbarnstillegg(
-        barnasAndeler: List<AndelTilkjentYtelse>,
-        barnasAktørerOgFødselsdatoer: List<Pair<Aktør, LocalDate>>
-    ): Tidslinje<BarnSinRettTilSmåbarnstillegg, Måned> {
-        val barnasAndelerTidslinjer = barnasAndeler.groupBy { it.aktør }.mapValues { AndelTilkjentYtelseTidslinje(it.value) }
-
-        val barnasAndelerUnder3ÅrTidslinje = barnasAndelerTidslinjer.map { (barnAktør, barnTidslinje) ->
-            val barnetsFødselsdato = barnasAktørerOgFødselsdatoer.find { it.first == barnAktør }?.second ?: throw Feil("Kan ikke beregne småbarnstillegg for et barn som ikke har fødselsdato.")
-
-            val under3ÅrTidslinje = erUnder3ÅrTidslinje(barnetsFødselsdato)
-
-            barnTidslinje.beskjærEtter(under3ÅrTidslinje)
-        }
-
-        return barnasAndelerUnder3ÅrTidslinje.kombinerUtenNull { kombinerBarnasTidslinjerTilUnder3ÅrResultat(it) }
     }
 
     fun lagPerioderMedBarnSomGirRettTilSmåbarnstilleggGammel(
